int64

+++ link_bot_pycommon/src/link_bot_pycommon/grid_utils.py
@@ -41,7 +41,7 @@ def lookup_points_in_vg(state_points, local_env, res, local_origin_point, batch_
     vg_indices = batch_point_to_idx(state_points,
                                     tf.expand_dims(res, axis=1),
                                     tf.expand_dims(local_origin_point, axis=1))
-    batch_indices = tf.tile(tf.range(batch_size)[:, None, None], [1, n_points, 1])
+    batch_indices = tf.tile(tf.range(batch_size, dtype=tf.int64)[:, None, None], [1, n_points, 1])
     batch_and_vg_indices = tf.concat([batch_indices, vg_indices], axis=-1)
     occupancy_at_state_points = tf.gather_nd(local_env, batch_and_vg_indices)
     return occupancy_at_state_points
